requirements.txt
torch
transformers
datasets
Pillow
requests
timm
flash_attn
einops